+++ assets/src/blocks/godam-player/edit-common-settings.js
@@ -24,8 +24,8 @@ const options = [
  * @return {WPElement} The video settings component.
  */
 const VideoSettings = ( { setAttributes, attributes, isInsideQueryLoop = false } ) => {
-	const { autoplay, controls, loop, muted, preload } =
-		attributes;
+	const { autoplay, controls, loop, muted, preload, showShareButton, engagements } =
+	attributes;
 
 	// Show a specific help for autoplay setting.
 	const getAutoplayHelp = useMemo( () => {
@@ -45,6 +45,14 @@ const VideoSettings = ( { setAttributes, attributes, isInsideQueryLoop = false }
 		return null;
 	}, [ autoplay, muted ] );
 
+	const getShowShareButtonHelp = useMemo( () => {
+		if ( ! showShareButton ) {
+			return __( 'Removes the share button from the video player.', 'godam' );
+		}
+
+		return null;
+	}, [ showShareButton ] );
+
 	const toggleFactory = useMemo( () => {
 		const toggleAttribute = ( attribute ) => {
 			return ( newValue ) => {
@@ -57,6 +65,8 @@ const VideoSettings = ( { setAttributes, attributes, isInsideQueryLoop = false }
 			loop: toggleAttribute( 'loop' ),
 			muted: toggleAttribute( 'muted' ),
 			controls: toggleAttribute( 'controls' ),
+			engagements: toggleAttribute( 'engagements' ),
+			showShareButton: toggleAttribute( 'showShareButton' ),
 		};
 	}, [ setAttributes ] );
 
@@ -100,6 +110,13 @@ const VideoSettings = ( { setAttributes, attributes, isInsideQueryLoop = false }
 				onChange={ toggleFactory.controls }
 				checked={ !! controls }
 			/>
+			<ToggleControl
+				__nextHasNoMarginBottom
+				label={ __( 'Share Button', 'godam' ) }
+				onChange={ toggleFactory.showShareButton }
+				checked={ !! showShareButton }
+				help={ getShowShareButtonHelp }
+			/>
 			{ ! isInsideQueryLoop && (
 				<SelectControl
 					__next40pxDefaultSize
@@ -111,6 +128,13 @@ const VideoSettings = ( { setAttributes, attributes, isInsideQueryLoop = false }
 					hideCancelButton
 				/>
 			) }
+			<ToggleControl
+				__nextHasNoMarginBottom
+				label={ __( 'Enable Likes & Comments', 'godam' ) }
+				onChange={ toggleFactory.engagements }
+				checked={ !! engagements }
+				help={ __( 'Engagement will only be visible for transcoded videos', 'godam' ) }
+			/>
 		</>
 	);
 };