+++ inc/helpers/autoloader.php
@@ -70,8 +70,9 @@ function autoloader( $file_resource = '' ) {
 			case 'elementor-controls': // phpcs:ignore
 			case 'wpforms': // phpcs:ignore
 			case 'ninja-forms': // phpcs:ignore
+			case 'everest-forms': // phpcs:ignore
 			case 'metform': // phpcs:ignore
-
+			case 'lifter-lms': // phpcs:ignore
 				/**
 				 * If there is class name provided for specific directory then load that.
 				 * otherwise find in inc/ directory.