+++ pages/godam/utils/index.js
@@ -25,7 +25,7 @@ export const isOnStarterPlan = window?.userData?.userApiData?.active_plan === 'S
 /**
  * GODAM API base URL
  */
-export const GODAM_API_BASE = 'https://godam.io';
+export const GODAM_API_BASE = window?.godamSettings?.apiBase || '';
 
 /**
  * Get max upload size in bytes.