+++ pages/media-library/components/modal/FolderCreationModal.jsx
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect, useRef, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 /**
@@ -30,6 +30,11 @@ const FolderCreationModal = () => {
 	const isOpen = useSelector( ( state ) => state.FolderReducer.modals.folderCreation );
 	const selectedFolder = useSelector( ( state ) => state.FolderReducer.currentContextMenuFolder );
 
+	const allFolders = useSelector( ( state ) => state.FolderReducer.folders );
+	const currentFolder = useMemo( () => {
+		return allFolders.find( ( folder ) => folder.id === selectedFolder?.id );
+	}, [ allFolders, selectedFolder ] );
+
 	useEffect( () => {
 		if ( isOpen ) {
 			setFolderName( '' );
@@ -52,6 +57,10 @@ const FolderCreationModal = () => {
 			let parent = selectedFolder?.id;
 
 			if ( selectedFolder?.id === -1 || ! selectedFolder?.id ) {
+				parent = 0;
+			}
+
+			if ( currentFolder?.meta?.locked ) {
 				parent = 0;
 			}