+++ webpack.config.js
@@ -142,13 +142,6 @@ const gfGodamRecorderEditorJS = {
 	},
 };
 
-const gfEntryDetailJS = {
-	...sharedConfig,
-	entry: {
-		'gf-entry-detail': path.resolve( process.cwd(), 'assets', 'src', 'js', 'gravity-form', 'gf-entry-detail.js' ),
-	},
-};
-
 const wpFormsGodamRecorderEditorJS = {
 	...sharedConfig,
 	entry: {
@@ -192,6 +185,44 @@ const fluentForms = {
 	},
 };
 
+const everestForms = {
+	...sharedConfig,
+	entry: {
+		everestforms: path.resolve( process.cwd(), 'assets', 'src', 'js', 'everestforms', 'index.js' ),
+	},
+};
+
+const godamPlayerSDK = {
+	...sharedConfig,
+	entry: {
+		godamPlayerSDK: path.resolve( process.cwd(), 'assets', 'src', 'js', 'godam-player', 'godam-player-sdk.js' ),
+		'godam-player-sdk': path.resolve( process.cwd(), 'assets', 'src', 'js', 'godam-player', 'godam-player-sdk.js' ),
+	},
+};
+
+const lifterLMSBlock = {
+	...sharedConfig,
+	entry: {
+		lifterLMSBlock: path.resolve( process.cwd(), 'assets', 'src', 'js', 'lifterlms', 'block.js' ),
+		'godam-lifterlms-block': path.resolve( process.cwd(), 'assets', 'src', 'js', 'lifterlms', 'block.js' ),
+	},
+};
+
+const lifterLMSEmbed = {
+	...sharedConfig,
+	entry: {
+		lifterLMSEmbed: path.resolve( process.cwd(), 'assets', 'src', 'js', 'lifterlms', 'embed.js' ),
+		'godam-lifterlms-embed': path.resolve( process.cwd(), 'assets', 'src', 'js', 'lifterlms', 'embed.js' ),
+	},
+};
+
+const ninjaForms = {
+	...sharedConfig,
+	entry: {
+		'ninja-forms': path.resolve( process.cwd(), 'assets', 'src', 'js', 'ninja-forms', 'index.js' ),
+	},
+};
+
 // Define the `pages` directory
 const pagesDir = path.resolve( __dirname, './pages' );
 
@@ -232,7 +263,7 @@ const pages = {
 				use: [ 'style-loader', 'css-loader', 'postcss-loader', 'sass-loader' ],
 			},
 			{
-				test: /\.(png|jpg|jpeg|gif|svg)$/, // Handle image files
+				test: /\.(png|jpg|jpeg|gif|svg|webp)$/, // Handle image files
 				use: [
 					{
 						loader: 'file-loader',
@@ -267,7 +298,6 @@ module.exports = [
 	deactivationJS,
 	godamGallery,
 	gfGodamRecorderEditorJS,
-	gfEntryDetailJS,
 	wpFormsGodamRecorderEditorJS,
 	jetpackFormJS,
 	styles, // Do not remove this.
@@ -276,4 +306,9 @@ module.exports = [
 	elementorEditorJS,
 	godamRecorder,
 	fluentForms,
+	everestForms,
+	godamPlayerSDK,
+	lifterLMSBlock,
+	lifterLMSEmbed,
+	ninjaForms,
 ];